Handle race condition that fails some upload failure retries [MITM-842]
There's a race condition that happens during retries -

onError() --> notifies app
onCompleted() --> is called after onError() and this is when UploadManager in rn-upload removes the entry from map.

If the app retries before onCompleted() is called, then the retry fails due to UploadManager thinking it's a duplicate upload (since the map still has the upload task).

So, to fix this, this PR removes the task from the map onError() in UploadManager.